+++ app/upgrades/v16/upgrades.go
@@ -7,6 +7,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
+	"github.com/evmos/evmos/v16/precompiles/bech32"
 	osmosisoutpost "github.com/evmos/evmos/v16/precompiles/outposts/osmosis"
 	strideoutpost "github.com/evmos/evmos/v16/precompiles/outposts/stride"
 	"github.com/evmos/evmos/v16/precompiles/p256"
@@ -26,11 +27,11 @@ func CreateUpgradeHandler(
 	return func(ctx sdk.Context, _ upgradetypes.Plan, vm module.VersionMap) (module.VersionMap, error) {
 		logger := ctx.Logger().With("upgrade", UpgradeName)
 
-		// enable secp256r1 precompile on testnet
+		// enable secp256r1 and bech32 precompile on testnet
 		if utils.IsTestnet(ctx.ChainID()) {
 			p256Address := p256.Precompile{}.Address()
-
-			if err := ek.EnablePrecompiles(ctx, p256Address); err != nil {
+			bech32Address := bech32.Precompile{}.Address()
+			if err := ek.EnablePrecompiles(ctx, p256Address, bech32Address); err != nil {
 				logger.Error("failed to enable precompiles", "error", err.Error())
 			}
 		}